requirements.txt
discord.py[voice]==1.3.4
google-cloud-texttospeech==2.1.0
